[SYMFONY 3.4] Fix update commands breaking after the container was cleared
Same problem as in cache clear admin action: when the container cache is
cleared and code which runs after the clear tries to fetch a service
from the container, it breaks as files might not be available anymore.

To resolve this, we explicitely exit() after the update commands and
also explicitely unregister any event listeners on the terminate event.